tools/urban-plot-analysis/requirements.txt
# Urban Plot Analysis - Dependencies

# API Clients
requests>=2.31.0
anthropic>=0.40.0

# Data Processing
pandas>=2.0.0
geopy>=2.4.0

# Image Processing
Pillow>=10.0.0

# Environment Configuration
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Type hints
typing-extensions>=4.8.0
